refactor: split generator logic
